+++ service/controller/resource/machinedeploymentstatus/resource.go
@@ -14,6 +14,7 @@ import (
 
 	"github.com/giantswarm/cluster-operator/pkg/label"
 	"github.com/giantswarm/cluster-operator/service/controller/key"
+	"github.com/giantswarm/cluster-operator/service/internal/basedomain"
 	"github.com/giantswarm/cluster-operator/service/internal/nodecount"
 	"github.com/giantswarm/cluster-operator/service/internal/tenantclient"
 )
@@ -74,6 +75,12 @@ func (r *Resource) ensure(ctx context.Context, obj interface{}) error {
 	workerCount, err := r.nodeCount.WorkerCount(ctx, cr)
 	if tenantclient.IsNotAvailable(err) {
 		r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("not getting worker nodes for tenant cluster %#q", key.ClusterID(cr)))
+		r.logger.LogCtx(ctx, "level", "debug", "message", "canceling resource")
+		resourcecanceledcontext.SetCanceled(ctx)
+		return nil
+	} else if basedomain.IsNotFound(err) {
+		// in case of a cluster deletion AWSCluster CR does not exist anymore, handle basedomain error gracefully
+		r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("not getting basedomain for tenant client %#q", key.ClusterID(cr)))
 		r.logger.LogCtx(ctx, "level", "debug", "message", "canceling resource")
 		resourcecanceledcontext.SetCanceled(ctx)
 		return nil